Clippy

+++ packages/morsels_common/src/bitmap.rs
@@ -7,7 +7,7 @@ pub fn check(vec: &[u8], at: usize) -> bool {
 
 #[inline(always)]
 pub fn set(vec: &mut [u8], at: usize) {
-    let byte_num = at / 8 as usize;
+    let byte_num = at / 8_usize;
     vec[byte_num] |= 1_u8 << (at % 8) as u8;
 }
